refactor: split predicate.go file
